# src/docs/swagger.yaml
definitions:
  dto.CreateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      title:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - description
    - status
    - title
    type: object
  dto.LoginByUsernameRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterUserByUsernameRequest:
    properties:
      email:
        minLength: 6
        type: string
      firstName:
        minLength: 3
        type: string
      lastName:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
  dto.UpdateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      title:
        maxLength: 255
        minLength: 1
        type: string
    type: object
  helpers.BaseHTTPResponse:
    properties:
      error: {}
      result: {}
      result_code:
        $ref: '#/definitions/helpers.ResultCode'
      success:
        type: boolean
      validation_errors:
        items:
          $ref: '#/definitions/validators.ValidationError'
        type: array
    type: object
  helpers.ResultCode:
    enum:
    - 0
    - 50001
    type: integer
    x-enum-varnames:
    - Success
    - CustomRecovery
  models.TaskStatus:
    enum:
    - pending
    - in_progress
    - completed
    type: string
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusInProgress
    - TaskStatusCompleted
  validators.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact:
    email: mahdipeydai@gmail.com
    name: Mahdi Peydai
  title: TaskManager - Go
  version: "0.1"
paths:
  /v1/health:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      security:
      - AuthBearer: []
  /v1/tasks:
    get:
      description: Get paginated tasks with optional status and assignee filters.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Task status
        in: query
        name: status
        type: string
      - description: Assignee ID
        in: query
        name: assignee_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      security:
      - AuthBearer: []
      summary: Get Tasks
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Create a new task.
      parameters:
      - description: Create task
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      security:
      - AuthBearer: []
      summary: Create Task
      tags:
      - Tasks
  /v1/tasks/{id}:
    delete:
      description: Soft delete a task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      security:
      - AuthBearer: []
      summary: Delete Task
      tags:
      - Tasks
    get:
      description: Get a task by ID.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      security:
      - AuthBearer: []
      summary: Get Task
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update an existing task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update task
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      security:
      - AuthBearer: []
      summary: Update Task
      tags:
      - Tasks
  /v1/users/login:
    post:
      consumes:
      - application/json
      description: Login By Username
      parameters:
      - description: Login by username
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      summary: Login By Username
      tags:
      - Users
  /v1/users/refresh-token:
    post:
      consumes:
      - application/json
      description: Generates a new access token and refresh token using a valid refresh
        token.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      summary: Refresh access token
      tags:
      - Users
  /v1/users/register:
    post:
      consumes:
      - application/json
      description: Register By Username
      parameters:
      - description: Register by username
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "404":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
        "500":
          description: Failed
          schema:
            $ref: '#/definitions/helpers.BaseHTTPResponse'
      summary: Register By Username
      tags:
      - Users
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"